requirements.txt
langchain-community
openai
langgraph
python-dotenv
langgraph
pinecone
pymupdf
tiktoken
